Factor out common section header code for lists.
