+++ src/read/rnglists.rs
@@ -1,11 +1,12 @@
 use crate::common::{
-    DebugAddrBase, DebugAddrIndex, DebugRngListsBase, DebugRngListsIndex, Encoding, Format,
+    DebugAddrBase, DebugAddrIndex, DebugRngListsBase, DebugRngListsIndex, Encoding,
     RangeListsOffset, SectionId,
 };
 use crate::constants;
 use crate::endianity::Endianity;
 use crate::read::{
-    DebugAddr, EndianSlice, Error, Reader, ReaderOffset, ReaderOffsetId, Result, Section,
+    lists::ListsHeader, DebugAddr, EndianSlice, Error, Reader, ReaderOffset, ReaderOffsetId,
+    Result, Section,
 };
 
 /// The raw contents of the `.debug_ranges` section.
@@ -100,63 +101,8 @@ impl<R> From<R> for DebugRngLists<R> {
     }
 }
 
-#[derive(Debug, Clone, Copy)]
-struct RngListsHeader {
-    encoding: Encoding,
-    offset_entry_count: u32,
-}
-
-impl Default for RngListsHeader {
-    fn default() -> Self {
-        RngListsHeader {
-            encoding: Encoding {
-                format: Format::Dwarf32,
-                version: 5,
-                address_size: 0,
-            },
-            offset_entry_count: 0,
-        }
-    }
-}
-
-impl RngListsHeader {
-    /// Return the serialized size of the table header.
-    #[allow(dead_code)]
-    #[inline]
-    fn size(self) -> u8 {
-        // initial_length + version + address_size + segment_selector_size + offset_entry_count
-        self.encoding.format.initial_length_size() + 2 + 1 + 1 + 4
-    }
-}
-
-// TODO: add an iterator over headers in the .debug_rnglists section
-#[allow(dead_code)]
-fn parse_header<R: Reader>(input: &mut R) -> Result<RngListsHeader> {
-    let (length, format) = input.read_initial_length()?;
-    input.truncate(length)?;
-
-    let version = input.read_u16()?;
-    if version != 5 {
-        return Err(Error::UnknownVersion(u64::from(version)));
-    }
-
-    let address_size = input.read_u8()?;
-    let segment_selector_size = input.read_u8()?;
-    if segment_selector_size != 0 {
-        return Err(Error::UnsupportedSegmentSize);
-    }
-    let offset_entry_count = input.read_u32()?;
-
-    let encoding = Encoding {
-        format,
-        version,
-        address_size,
-    };
-    Ok(RngListsHeader {
-        encoding,
-        offset_entry_count,
-    })
-}
+#[allow(unused)]
+pub(crate) type RngListsHeader = ListsHeader;
 
 /// The DWARF data found in `.debug_ranges` and `.debug_rnglists` sections.
 #[derive(Debug, Default, Clone, Copy)]
@@ -600,6 +546,7 @@ impl Range {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::common::Format;
     use crate::endianity::LittleEndian;
     use crate::test_util::GimliSectionMethods;
     use test_assembler::{Endian, Label, LabelMaker, Section};
